@@ -4,9 +4,8 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
-  flutter: ">=1.17.0"
-
+  sdk: '>=3.0.0 <4.0.0'
+  
 dependencies:
   flutter:
     sdk: flutter
